completions/gpt.py
import os

from openai import OpenAI
from dotenv import load_dotenv

load_dotenv()

key = os.getenv("OPENAI_API_KEY")
organization = os.getenv("OPENAI_ORGANIZATION")

client = OpenAI(api_key=key, base_url="https://api.openai.com/v1", organization="org-62CllBpIXV3KJBB9kJMO2bAs")


chat_completion = client.chat.completions.create(
    messages=[
        {
            "role": "user",
            "content": "Say this is a test",
        }
    ],
    model="gpt-4o-mini",
)

print(chat_completion)
