requirements.txt
anthropic
cohere
datasets
matplotlib
numpy
openai
python-dotenv
torch